+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/Spaceship.java
@@ -8,7 +8,9 @@ import java.util.LinkedList;
 
 public class Spaceship extends SpaceEntity {
 
+    public static final int DURABILITY = 100;
+
     public Spaceship(LinkedList<Entity> missileSubscriber, float x, float y) {
-        super(EntityType.SPACESHIP, null, null, EntityTurn.OUTCOMING, new Shooting(missileSubscriber, MissileType.LASER), EntityShapes.SPACESHIP_SHAPE, GameImage.SPACESHIP, x, y, 0.075f, 100, 0);
+        super(EntityType.SPACESHIP, null, null, EntityTurn.OUTCOMING, new Shooting(missileSubscriber, MissileType.LASER), EntityShapes.SPACESHIP_SHAPE, GameImage.SPACESHIP, x, y, 0.075f, DURABILITY, 0);
     }
 }