+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.20.0] - 2023-09-11
+## [0.21.0] - 2023-08-12
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): `GameBar` component.
+- [@mjkuranda](https://github.com/mjkuranda): `LongGameBar` component.
+- [@mjkuranda](https://github.com/mjkuranda): Time and durability as game bars.
+- [@mjkuranda](https://github.com/mjkuranda): Score, nickname and missile count as text on the screen.
+- [@mjkuranda](https://github.com/mjkuranda): Empty input for your name gives `Noname` player name.
+- [@mjkuranda](https://github.com/mjkuranda): Info rendering - press `H` to toggle its.
+
+## [0.20.0] - 2023-08-11
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Game time up to 5 minutes.
 - [@mjkuranda](https://github.com/mjkuranda): Spawning a new entity probability depending on left game time.